+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- A diagnostic command-line-option to erase version history. Helpful for recovering when the version history is corrupt.
+
+### Fixed
+
+- Backing up data into another KeyValueDatabase could lead to inconsistencies if the backup was immediately used.
+
 ## [0.2.0] - 2021-08-13
 
 ### Fixed